requirements.txt
discord.py>=2.0.0
pyautogui>=0.9.53
opencv-python>=4.5.0
numpy>=1.19.0
pyperclip>=1.8.0
keyboard>=0.13.0
sounddevice>=0.4.0
scipy>=1.6.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0
requests>=2.25.0
pillow>=8.0.0
python-dotenv>=0.19.0